chore: updated observation map layer
